+++ src/components/plips-layout.js
@@ -2,8 +2,8 @@ import PropTypes from "prop-types";
 import React, { Component } from "react";
 
 import {
+  SafeAreaView,
   TouchableOpacity,
-  View,
 } from "react-native";
 
 import {
@@ -19,6 +19,10 @@ import MainTabView from "../components/main-tab-view";
 
 import styles from "../styles/plips-layout";
 
+import SearchModal from "./search-modal";
+
+import { getMainTabViewIndexByKey } from "./../models/main-tab-view";
+
 import {
   RemoteLinksType,
 } from "../prop-types";
@@ -27,6 +31,7 @@ export default class PlipsLayout extends Component {
   static propTypes = {
     allPlips: PropTypes.array,
     allPlipsNextPage: PropTypes.number,
+    currentMainTabViewTab: PropTypes.string,
     currentUser: PropTypes.object,
     favoritePlips: PropTypes.array,
     favoritePlipsNextPage: PropTypes.number,
@@ -40,30 +45,63 @@ export default class PlipsLayout extends Component {
     openMenu: PropTypes.func.isRequired,
     plipsSignInfo: PropTypes.object.isRequired,
     remoteLinks: RemoteLinksType,
+    searchTitle: PropTypes.string,
     signedPlips: PropTypes.array,
     signedPlipsNextPage: PropTypes.number,
     tabViewState: TabViewType,
     userLocationPlips: PropTypes.array,
     userLocationPlipsNextPage: PropTypes.number,
     userSignInfo: PropTypes.object.isRequired,
+    onClearSearch: PropTypes.func.isRequired,
     onFetchPlipsNextPage: PropTypes.func.isRequired,
     onGoToPlip: PropTypes.func.isRequired,
     onMainTabChange: PropTypes.func.isRequired,
     onOpenURL: PropTypes.func.isRequired,
     onRefresh: PropTypes.func.isRequired,
     onRetryPlips: PropTypes.func.isRequired,
+    onSearchPlip: PropTypes.func.isRequired,
     onShare: PropTypes.func.isRequired,
     onToggleFavorite: PropTypes.func.isRequired,
   }
 
+  state = {
+    isSearchModalVisible: false,
+  };
+
+  goToAllPlips = () => {
+    const { onMainTabChange } = this.props;
+
+    const index = getMainTabViewIndexByKey("allPlips")
+    onMainTabChange({ index });
+  };
+
+  onToggleSearchModal = () => {
+    const { currentMainTabViewTab } = this.props;
+    const { isSearchModalVisible } = this.state;
+
+    if (!isSearchModalVisible && currentMainTabViewTab !== "allPlips") {
+      this.goToAllPlips();
+    }
+
+    this.setState(({ isSearchModalVisible }) => ({ isSearchModalVisible: !isSearchModalVisible }));
+  };
+
+  onClearSearch = () => {
+    const { onClearSearch } = this.props;
+
+    this.onToggleSearchModal();
+    onClearSearch();
+  }
+
   render() {
     return (
-      <View style={styles.full}>
+      <SafeAreaView style={styles.full}>
         <Layout>
-          {this.renderNavBar()}
+          { this.renderNavBar() }
           <MainTabView {...this.props}/>
+          { this.renderSearchModal() }
         </Layout>
-      </View>
+      </SafeAreaView>
     );
   }
 
@@ -96,11 +134,24 @@ export default class PlipsLayout extends Component {
     );
   }
 
-  // TODO: Add search filter
+  renderSearchModal() {
+    const { isSearchModalVisible } = this.state;
+    const { searchTitle, onSearchPlip } = this.props;
+
+    return (
+      <SearchModal
+        isVisible={isSearchModalVisible}
+        onToggleModal={this.onToggleSearchModal}
+        onSearch={onSearchPlip}
+        onClearSearch={this.onClearSearch}
+        searchTitle={searchTitle} />
+    );
+  }
+
   renderSearchButton() {
     return (
       <TouchableOpacity
-        onPress={this.onOpenMenu}
+        onPress={this.onToggleSearchModal}
       >
         <Icon
           name="search"